requirements.txt
Django==1.10.6
django-nested-admin==3.0.12
django-polymorphic==1.1
ipdb==0.10.1
ipython==5.3.0
Markdown==2.6.8
psycopg2==2.6.2
